server.js and productcontroller file modified

+++ backend/controllers/product.controller.js
@@ -46,12 +46,16 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) =>{
     const{ id } = req.params;
+
+    if( !mongoose.Types.ObjectId.isValid(id)){
+      return res.status(404).json({success: false, message: "Invalid product ID"});
+    }
     
     try{
       await Product.findByIdAndDelete(id);
       res.status(200).json({sucess: true, message: "Product successfully deleted" });
     } catch (error){
       console.log("error in deleting product:", error.message);
-      res.status(404).json({success: false, message: "Product not found"});
+      res.status(500).json({success: false, message: "Server Error"});
     }
   };
